trying to figure out what init() needs to work

+++ VL53L0X.h
@@ -9,11 +9,11 @@
 
 // DEFINE ANY MACROS?
 
-void vl53l0x_write16(uint16_t reg, uint16_t val, uint8_t addr);
-void vl53l0x_write8(uint16_t reg, uint8_t val, uint8_t addr);
+void vl53l0x_write16(uint8_t reg, uint16_t val, uint8_t addr);
+void vl53l0x_write8(uint8_t reg, uint8_t val, uint8_t addr);
 
-uint16_t vl53l0x_read16(uint16_t reg, uint8_t addr);
-uint8_t vl53l0x_read8(uint16_t reg, uint8_t addr);
+uint16_t vl53l0x_read16(uint8_t reg, uint8_t addr);
+uint8_t vl53l0x_read8(uint8_t reg, uint8_t addr);
 
 void vl53l0x_init(void);
 void vl53l0x_setaddress(uint16_t newAddr);
